+++ ppdbSma/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Lenovo\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
+user.properties.file=C:\\Users\\Fariz\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
