+++ ppdbSma/src/gui/FrameLogin.java
@@ -5,6 +5,8 @@
  */
 package gui;
 
+import logic.LFLogin;
+import logic.LFSiswa;
 import logic.SqlConnection;
 
 /**
@@ -36,12 +38,7 @@ public class FrameLogin extends javax.swing.JFrame {
         fieldPassword = new javax.swing.JPasswordField();
         fieldAlamatDb = new javax.swing.JTextField();
         jLabel3 = new javax.swing.JLabel();
-        jLabel4 = new javax.swing.JLabel();
         tombolOk1 = new javax.swing.JButton();
-        fieldPasswordDb = new javax.swing.JPasswordField();
-        jLabel5 = new javax.swing.JLabel();
-        fieldUserDb = new javax.swing.JTextField();
-        tombolSiswa = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -51,7 +48,7 @@ public class FrameLogin extends javax.swing.JFrame {
             }
         });
 
-        jLabel1.setText("nisn");
+        jLabel1.setText("username");
 
         jLabel2.setText("password");
 
@@ -63,33 +60,10 @@ public class FrameLogin extends javax.swing.JFrame {
 
         jLabel3.setText("alamat_db");
 
-        jLabel4.setText("user_db");
-
-        tombolOk1.setText("OK");
+        tombolOk1.setText("Login");
         tombolOk1.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 tombolOk1ActionPerformed(evt);
-            }
-        });
-
-        fieldPasswordDb.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                fieldPasswordDbActionPerformed(evt);
-            }
-        });
-
-        jLabel5.setText("pass_db");
-
-        fieldUserDb.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                fieldUserDbActionPerformed(evt);
-            }
-        });
-
-        tombolSiswa.setText("Siswa");
-        tombolSiswa.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                tombolSiswaActionPerformed(evt);
             }
         });
 
@@ -98,34 +72,25 @@ public class FrameLogin extends javax.swing.JFrame {
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
+                .addGap(33, 33, 33)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jLabel1)
+                        .addComponent(jLabel2, javax.swing.GroupLayout.Alignment.TRAILING))
                     .addGroup(layout.createSequentialGroup()
-                        .addGap(33, 33, 33)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel1)
-                            .addComponent(jLabel2, javax.swing.GroupLayout.Alignment.TRAILING))
-                        .addGap(32, 32, 32)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(fieldNisn, javax.swing.GroupLayout.DEFAULT_SIZE, 104, Short.MAX_VALUE)
-                            .addComponent(fieldPassword)))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(101, 101, 101)
-                        .addComponent(tombolSiswa)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 51, Short.MAX_VALUE)
+                        .addGap(2, 2, 2)
+                        .addComponent(jLabel3)))
+                .addGap(27, 27, 27)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addComponent(jLabel3)
-                        .addComponent(jLabel4, javax.swing.GroupLayout.Alignment.TRAILING))
-                    .addComponent(jLabel5))
-                .addGap(32, 32, 32)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                    .addComponent(fieldPasswordDb, javax.swing.GroupLayout.DEFAULT_SIZE, 193, Short.MAX_VALUE)
-                    .addComponent(fieldAlamatDb, javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(fieldUserDb, javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(tombolOk1)
-                        .addGap(79, 79, 79)))
-                .addGap(60, 60, 60))
+                    .addComponent(fieldAlamatDb, javax.swing.GroupLayout.DEFAULT_SIZE, 359, Short.MAX_VALUE)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                        .addComponent(fieldNisn, javax.swing.GroupLayout.DEFAULT_SIZE, 104, Short.MAX_VALUE)
+                        .addComponent(fieldPassword)))
+                .addGap(139, 139, 139))
+            .addGroup(layout.createSequentialGroup()
+                .addGap(94, 94, 94)
+                .addComponent(tombolOk1)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -133,20 +98,10 @@ public class FrameLogin extends javax.swing.JFrame {
                 .addGap(39, 39, 39)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(layout.createSequentialGroup()
+                        .addGap(70, 70, 70)
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                             .addComponent(fieldAlamatDb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel3))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel4)
-                            .addComponent(fieldUserDb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(layout.createSequentialGroup()
-                                .addGap(16, 16, 16)
-                                .addComponent(jLabel5))
-                            .addGroup(layout.createSequentialGroup()
-                                .addGap(18, 18, 18)
-                                .addComponent(fieldPasswordDb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                            .addComponent(jLabel3)))
                     .addGroup(layout.createSequentialGroup()
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                             .addComponent(fieldNisn, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
@@ -154,12 +109,10 @@ public class FrameLogin extends javax.swing.JFrame {
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                             .addComponent(jLabel2)
-                            .addComponent(fieldPassword, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(18, 18, 18)
-                        .addComponent(tombolSiswa)))
-                .addGap(25, 25, 25)
+                            .addComponent(fieldPassword, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                .addGap(27, 27, 27)
                 .addComponent(tombolOk1)
-                .addContainerGap(121, Short.MAX_VALUE))
+                .addContainerGap(127, Short.MAX_VALUE))
         );
 
         pack();
@@ -169,36 +122,28 @@ public class FrameLogin extends javax.swing.JFrame {
         // TODO add your handling code here:
     }//GEN-LAST:event_fieldNisnActionPerformed
 
-    private void fieldAlamatDbActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_fieldAlamatDbActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_fieldAlamatDbActionPerformed
-
     private void tombolOk1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_tombolOk1ActionPerformed
         // TODO add your handling code here:
         String alamat = fieldAlamatDb.getText();
-        String user = fieldUserDb.getText();
-        String pass = fieldPasswordDb.getText();
-        SqlConnection.setLogin("jdbc:mysql://" + alamat, user, pass);
-        FrameAdmin frameAdmin = new FrameAdmin();
-        frameAdmin.setVisible(true);
-        frameAdmin.setDefaultCloseOperation(DISPOSE_ON_CLOSE);
+        SqlConnection.setLogin("jdbc:mysql://localhost:3306/ppdbSman" + alamat, "root", "");
+        String user = fieldNisn.getText();
+        String pass = fieldPassword.getText();
+        String status = LFLogin.login(user, pass);
+        if(status.equals("admin")){
+            FrameAdmin frameAdmin = new FrameAdmin();
+            frameAdmin.setVisible(true);
+        } else if(status.equals("gagal")){
+            
+        } else{
+            LFSiswa.nisn = user;
+            FrameSiswa frameSiswa = new FrameSiswa();
+            frameSiswa.setVisible(true);
+        }
     }//GEN-LAST:event_tombolOk1ActionPerformed
 
-    private void fieldUserDbActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_fieldUserDbActionPerformed
+    private void fieldAlamatDbActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_fieldAlamatDbActionPerformed
         // TODO add your handling code here:
-    }//GEN-LAST:event_fieldUserDbActionPerformed
-
-    private void tombolSiswaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_tombolSiswaActionPerformed
-        // TODO add your handling code here:
-        FrameSiswaPilih formSiswaPilih = new FrameSiswaPilih();
-        formSiswaPilih.setVisible(true);
-        fieldNisn.getText();
-        
-    }//GEN-LAST:event_tombolSiswaActionPerformed
-
-    private void fieldPasswordDbActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_fieldPasswordDbActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_fieldPasswordDbActionPerformed
+    }//GEN-LAST:event_fieldAlamatDbActionPerformed
 
     /**
      * @param args the command line arguments
@@ -239,14 +184,9 @@ public class FrameLogin extends javax.swing.JFrame {
     private javax.swing.JTextField fieldAlamatDb;
     private javax.swing.JTextField fieldNisn;
     private javax.swing.JPasswordField fieldPassword;
-    private javax.swing.JPasswordField fieldPasswordDb;
-    private javax.swing.JTextField fieldUserDb;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
     private javax.swing.JButton tombolOk1;
-    private javax.swing.JButton tombolSiswa;
     // End of variables declaration//GEN-END:variables
 }
