--- ppdbSma/src/gui/FrameSiswaPilih.java
@@ -1,229 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gui;
-
-import java.sql.Connection;
-import java.util.ArrayList;
-import java.util.List;
-import logic.SqlConnection;
-
-/**
- *
- * @author Lenovo
- */
-public class FrameSiswaPilih extends javax.swing.JFrame {
-
-    /**
-     * Creates new form FormSiswaPilih
-     */
-    List<String> arraySekolah = new ArrayList<>();
-    public void imporDbSekolah(){
-        try {
-//            SqlConnection.setLogin("jdbc:mysql://localhost:9993/ppdbSmanDepok", "herbiejago", "sudo");
-            String sql = "select * from sekolah";
-            java.sql.Connection conn = (Connection)SqlConnection.SqlConnectionDB();
-            java.sql.Statement stm = conn.createStatement();
-            java.sql.ResultSet res = stm.executeQuery(sql);
-            while(res.next()) {
-                arraySekolah.add(res.getString(2));
-                System.out.println(res.getString(2));
-            }
-        } catch(Exception e) {
-
-        }
-    }
-    public FrameSiswaPilih() {
-        imporDbSekolah();
-        initComponents();
-        
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        TFNisn = new javax.swing.JTextField();
-        LNisn = new javax.swing.JLabel();
-        LNama = new javax.swing.JLabel();
-        TFNama = new javax.swing.JTextField();
-        LPil1 = new javax.swing.JLabel();
-        LPil2 = new javax.swing.JLabel();
-        LPil3 = new javax.swing.JLabel();
-        CBPil2 = new javax.swing.JComboBox<>();
-        CBPil3 = new javax.swing.JComboBox<>();
-        CBPil1 = new javax.swing.JComboBox<>();
-        tombolUbah = new javax.swing.JButton();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        textPane1 = new javax.swing.JTextPane();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        LNisn.setText("NISN");
-
-        LNama.setText("Nama");
-
-        TFNama.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                TFNamaActionPerformed(evt);
-            }
-        });
-
-        LPil1.setText("Pilihan 1");
-
-        LPil2.setText("Pilihan 2");
-
-        LPil3.setText("Pilihan 3");
-
-        CBPil2.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
-
-        CBPil3.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
-
-        CBPil1.setModel(new javax.swing.DefaultComboBoxModel(arraySekolah.toArray()
-        ));
-        CBPil1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                CBPil1ActionPerformed(evt);
-            }
-        });
-
-        tombolUbah.setText("Ubah");
-        tombolUbah.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                tombolUbahActionPerformed(evt);
-            }
-        });
-
-        jScrollPane1.setViewportView(textPane1);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(LNama)
-                    .addComponent(LNisn)
-                    .addComponent(LPil1)
-                    .addComponent(LPil2)
-                    .addComponent(LPil3))
-                .addGap(41, 41, 41)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(tombolUbah)
-                    .addComponent(CBPil1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                        .addComponent(CBPil3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(CBPil2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(TFNisn)
-                        .addComponent(TFNama)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 120, Short.MAX_VALUE)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 202, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(182, 182, 182))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(37, 37, 37)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 239, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(TFNama, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(LNama))
-                        .addGap(20, 20, 20)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(TFNisn, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(LNisn))
-                        .addGap(18, 18, 18)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(LPil1)
-                            .addComponent(CBPil1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(18, 18, 18)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(LPil2)
-                            .addComponent(CBPil2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(LPil3)
-                            .addComponent(CBPil3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(18, 18, 18)
-                        .addComponent(tombolUbah)))
-                .addContainerGap(24, Short.MAX_VALUE))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void TFNamaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_TFNamaActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_TFNamaActionPerformed
-
-    private void CBPil1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_CBPil1ActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_CBPil1ActionPerformed
-
-    private void tombolUbahActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_tombolUbahActionPerformed
-        // TODO add your handling code here:
-        String pilihan1 = CBPil1.getSelectedItem().toString();
-        textPane1.setText(pilihan1);
-    }//GEN-LAST:event_tombolUbahActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-//    public static void main(String args[]) {
-//        /* Set the Nimbus look and feel */
-//        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-//        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-//         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-//         */
-//        try {
-//            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-//                if ("Nimbus".equals(info.getName())) {
-//                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-//                    break;
-//                }
-//            }
-//        } catch (ClassNotFoundException ex) {
-//            java.util.logging.Logger.getLogger(FormSiswaPilih.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (InstantiationException ex) {
-//            java.util.logging.Logger.getLogger(FormSiswaPilih.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (IllegalAccessException ex) {
-//            java.util.logging.Logger.getLogger(FormSiswaPilih.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-//            java.util.logging.Logger.getLogger(FormSiswaPilih.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-//        }
-//        //</editor-fold>
-//
-//        /* Create and display the form */
-//        java.awt.EventQueue.invokeLater(new Runnable() {
-//            public void run() {
-//                new FormSiswaPilih().setVisible(true);
-//            }
-//        });
-//    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JComboBox<String> CBPil1;
-    private javax.swing.JComboBox<String> CBPil2;
-    private javax.swing.JComboBox<String> CBPil3;
-    private javax.swing.JLabel LNama;
-    private javax.swing.JLabel LNisn;
-    private javax.swing.JLabel LPil1;
-    private javax.swing.JLabel LPil2;
-    private javax.swing.JLabel LPil3;
-    private javax.swing.JTextField TFNama;
-    private javax.swing.JTextField TFNisn;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTextPane textPane1;
-    private javax.swing.JButton tombolUbah;
-    // End of variables declaration//GEN-END:variables
-}
